Fixed a Solr bug that prevented errors on startup from being logged.
CoreContainer in Solr 4.10.4 eats Errors (not Exceptions) when a core
fails to load. This can make for very frustrating debugging especially
when running in a Hadoop environment. This is a replacement
implementation that is useful for debugging. It lives in the test
package so as not to interfere with normal execution. It can be copied
into your hadoop job manually if necessary to override the default
implementation that has poor logging.

This bug has been corrected in later versions of Solr.